desloppify: convention + cross-module architecture cleanup
- Drop the lone AdoptionCandidate re-export at the core crate root
  (kept the per-vendor structs in their own modules — no external
  consumer used the bare crate-root alias).
- Rename src/window.rs to src/main.rs to match the idiomatic Rust
  binary entry. Cargo.toml [[bin]] path and one HANDOFF.md ref
  updated.
- Rename claude_telemetry::TranscriptTail to ClaudeTranscriptTail
  so all four agent transcript types share the vendor-prefixed
  shape (ClaudeTranscriptTail / CopilotTranscriptTail /
  OpenCodeMessageTail / PiTranscriptTail).
- Move vendor-neutral telemetry primitives (SessionState,
  TelemetrySnapshot, FileTail, context_window_for_model) out of
  core::claude_telemetry into a new core::telemetry module.
  copilot/opencode/pi telemetry modules import from the new home;
  claude_telemetry re-exports for backward compat. Breaks the
  'Claude is the de-facto base' import coupling without changing
  any behaviour.
- Note the git_status -> Result<Option<...>> change (prior commit)
  also resolves the contract-coherence mismatch with is_dirty.

Resolves desloppify review findings adoption_candidate_reexport_asymmetry,
binary_entry_filename, telemetry_tail_naming_drift,
telemetry_shared_types_misplaced, and git_error_contract_mismatch.
